+++ backend/controllers/usersController.js
@@ -1,0 +1,3 @@
+export const user = (req, res) => {
+
+}
